requirements-production.txt
asgiref==3.4.1
Django==3.2.8
pytz==2021.3
sqlparse==0.4.2
Pillow==8.4.0
django-versatileimagefield==2.2
django-cleanup==5.2.0